add: export using Sheet.js

+++ frontend/app/shared/infrastructure/db/InvoiceRepositoryImpl.ts
@@ -8,7 +8,7 @@ export class InvoiceRepositoryImpl implements InvoiceRepository {
 
     async deleteInvoiceByCustomerId(customerId: number): Promise<void> {
         const invoices = await this.db.selectFrom('invoice').where('customerId', '=', customerId).select('invoiceId').execute()
-        if (invoices?.length > 0) {
+        if (invoices?.length > 0 && invoices[0].invoiceId) {
             await this.db.deleteFrom('invoiceLine').where('invoiceId', 'in', [...invoices.flatMap((v) => v.invoiceId)]).execute()
             await this.db.deleteFrom('invoice').where('customerId', '=', customerId).execute()
         }
